+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## 1.0.0
+## 0.1.0-beta.1
 
-- Initial version, created by Stagehand
+- **SubtitleProvider:** Simplifies fetching subtitle file data from multiple sources.
+- **SubtitleRepository:** Deals with the platform directly to get or download the required data and submit it to the provider.
+- **SubtitleParser:** Used to analyze and convert subtitle files into software objects that are viewable and usable.
+- **CustomSubtitleParser:** Customizable subtitle parser, for custom regexes.
+- **SubtitleRegexObject:** Responsible for providing the necessary expression for the purpose of decoding the content of subtitle files, has 4 children:
+    - **VttRegex:** WebVTT regexp.
+    - **SrtRegex:** SubRip regexp.
+    - **TtmlRegex:** TTML or DFXP regex
+    - **CustomRegex:** User define regexp.
+- **SubtitleObject:** Store the subtitle file data and its format type.
+- Initial version, created by [MuhmdHsn313](https://twitter.com/MuhmdHsn313)